address review: split test files, remove pull-requests:write
- Split ai_agent_test.go into Tier 0 (offline) and
  ai_agent_recording_test.go (Tier 1 recording tests)
- Remove unnecessary pull-requests: write permission from workflow
- Actions versions already updated to v6 by prior commit
