+++ .github/workflows/lint-ext-azure-ai-agents.yml
@@ -5,6 +5,7 @@ on:
     paths:
       - "cli/azd/extensions/azure.ai.agents/**"
       - "cli/azd/test/functional/ai_agent_test.go"
+      - "cli/azd/test/functional/ai_agent_recording_test.go"
       - "cli/azd/test/functional/testdata/recordings/Test_AIAgent_*"
       - ".github/workflows/lint-ext-azure-ai-agents.yml"
     branches: [main]
@@ -15,7 +16,6 @@ concurrency:
 
 permissions:
   contents: read
-  pull-requests: write
 
 jobs:
   lint:
